fix: update wrappers for latest ipc state (#44)
The TL;DR is that there are a few variable name changes
(like creditDebitRate -> blobCreditsPerByteBlock or receiver -> to),
new fields that were unaccounted for, and the way credit approvals were
stored has changed. There's also a fix for the BucketManager contract,
which had an extra CBOR field, causing bucket creation to fail.
The best way to understand the changes is the README diff since it shows
what the new fields look like compared to the older setup.

I also created a quick and dirty script for "integration" tests.
It's not part of CI or anything, but it runs through everything I
usually do manually, as described in the README + cast / forge.
If you have localnet running in the background, you should be able to
run sh test/scripts/wrapper_integration_test.sh. If anything fails,
the script will end early; else, it'll log All tests completed
successfully at the end.

Closes #43