feat: reimplement validator rewarder (#69)
This PR reimplements validator rewarder contract according to a fixed
token increments. Please checkout the discussion here for more details
about the current implementation:
https://github.com/recallnet/ipc/issues/151#issuecomment-2526230269


With these changes there won't be any "state" in the rewarder contract.
The overall architecture remains the same. That is the validator
rewarder is being called by the IPC when there is a new claim for
rewards. We assume that IPC guarantees that there won't be any "double
claiming". IPC notifies the validator rewarder with the checkpoint
height, validator's address and the number of blocks that _this_
validator has committed in the checkpoint.

Each checkpoint contains 600 blocks. Which means 200 new tokens (1 token
per 3 blocks) are minted in each checkpoint. We just calculate the
validator's share of these 200 new tokens i.e. the ratio of blocks that
they have committed to total blocks in the checkpoint. We mint this
amount to validator's address.

---------

Signed-off-by: avichalp <hi@avichalp.me>