fix: deploy script contract shadowed declaration warning (#90)
When the `ipc` localnet starts there is a warning that reads: 
```
Compiler run successful with warnings:
Warning (2519): This declaration shadows an existing declaration.
  --> script/Recall.s.sol:61:30:
   |
61 |     function setDefaultRoles(address proxy, address admin, address minter, address pauser) public {
   |                              ^^^^^^^^^^^^^
Note: The shadowed declaration is here:
  --> script/Recall.s.sol:18:5:
   |
18 |     function proxy() public view returns (address) {
   |     ^ (Relevant source part starts here and spans across multiple lines).
```

This is because the contract has a method named `proxy` so the argument
shadows the method and makes it unreachable inside the `setDefaultRoles`
method.

It's not a big problem, but it would be nice to eliminate the warning
log during the localnet startup.