fix: update rewarder implementation (#105)
This pull request updates `ValidatorRewarder` contract and related
scripts and tests. The changes focus on improving the accuracy of reward
calculations.

In the earlier implementation we mistakenly assumed that every bottom-up
checkpoint will always be 600 blocks (checkpoint period). This
assumption is wrong. Bottom up checkpoints in IPC can be shorter than
the checkpoint period. Sometimes when the max number of messages are
added the checkpoint is sent immediately.

These changes reads the actual number of blocks in the checkpoint and
use that number for reward calculation. For instance, if any checkpoint
contains only 300 blocks then the rewards claimed by the validator
should be: 100 * validator's share.

closes: https://github.com/recallnet/contracts/issues/100

---------

Signed-off-by: avichalp <hi@avichalp.me>